% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pytekt.R
\docType{package}
\name{pytekt-package}
\alias{pytekt}
\alias{pytekt-package}
\title{Pytekt Engine for R}
\description{
Provides native high-performance C++ numerical kernels for 1D reductions,
vector dot products, matrix multiplication, rolling window routines, and histogramming.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aqwel/pytekt}
  \item Report bugs at \url{https://github.com/aqwel/pytekt/issues}
}

}
\author{
\strong{Maintainer}: Aksel Aghajanyan \email{aksel@aqwel.ai}

Authors:
\itemize{
  \item Aksel Aghajanyan \email{aksel@aqwel.ai}
}

}
\keyword{internal}
