% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_argmax}
\alias{pytekt_argmax}
\title{Fast Argmax (1-indexed)}
\usage{
pytekt_argmax(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
1-based index of maximum element
}
\description{
Fast Argmax (1-indexed)
}
\examples{
pytekt_argmax(c(4.0, 1.0, 8.0, 2.0))
}
