% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_argmin}
\alias{pytekt_argmin}
\title{Fast Argmin (1-indexed)}
\usage{
pytekt_argmin(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
1-based index of minimum element
}
\description{
Fast Argmin (1-indexed)
}
\examples{
pytekt_argmin(c(4.0, 1.0, 8.0, 2.0))
}
