% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{pytekt_available}
\alias{pytekt_available}
\title{Check if Pytekt Python Module is Available}
\usage{
pytekt_available()
}
\value{
Logical TRUE if Python environment has `aion` or `pytekt` installed, FALSE otherwise.
}
\description{
Check if Pytekt Python Module is Available
}
\examples{
pytekt_available()
}
