% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_chunk_statistics}
\alias{pytekt_chunk_statistics}
\title{Fast Chunk Statistics}
\usage{
pytekt_chunk_statistics(x, chunk_size)
}
\arguments{
\item{x}{Numeric vector}

\item{chunk_size}{Size of each chunk}
}
\value{
A list with `mean`, `min`, `max`, and `chunk_size`
}
\description{
Fast Chunk Statistics
}
\examples{
pytekt_chunk_statistics(c(1.0, 2.0, 3.0, 4.0, 5.0, 6.0), chunk_size = 2)
}
