% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_cosine_similarity}
\alias{pytekt_cosine_similarity}
\title{Fast Cosine Similarity}
\usage{
pytekt_cosine_similarity(a, b)
}
\arguments{
\item{a}{First numeric vector}

\item{b}{Second numeric vector}
}
\value{
Cosine similarity in range [-1, 1]
}
\description{
Fast Cosine Similarity
}
\examples{
pytekt_cosine_similarity(c(1.0, 0.0), c(1.0, 1.0))
}
