% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_dot}
\alias{pytekt_dot}
\title{Fast Dot Product of Two Vectors}
\usage{
pytekt_dot(a, b)
}
\arguments{
\item{a}{First numeric vector}

\item{b}{Second numeric vector}
}
\value{
Inner dot product
}
\description{
Fast Dot Product of Two Vectors
}
\examples{
pytekt_dot(c(1.0, 2.0), c(3.0, 4.0))
}
