% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_euclidean_distance}
\alias{pytekt_euclidean_distance}
\title{Fast Euclidean Distance}
\usage{
pytekt_euclidean_distance(a, b)
}
\arguments{
\item{a}{First numeric vector}

\item{b}{Second numeric vector}
}
\value{
Euclidean distance between a and b
}
\description{
Fast Euclidean Distance
}
\examples{
pytekt_euclidean_distance(c(0.0, 0.0), c(3.0, 4.0))
}
