% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_histogram}
\alias{pytekt_histogram}
\title{Fast Streaming Histogram}
\usage{
pytekt_histogram(x, bins, lo, hi)
}
\arguments{
\item{x}{Numeric vector}

\item{bins}{Number of bins}

\item{lo}{Lower bound}

\item{hi}{Upper bound}
}
\value{
Integer bin counts
}
\description{
Fast Streaming Histogram
}
\examples{
pytekt_histogram(c(1.0, 2.0, 3.0, 4.0, 5.0), bins = 5, lo = 1.0, hi = 5.0)
}
