% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_matrix_multiply}
\alias{pytekt_matrix_multiply}
\title{Fast Matrix Multiplication (Row-Major flattened matrices)}
\usage{
pytekt_matrix_multiply(A, B, m, n, p)
}
\arguments{
\item{A}{Flattened matrix A (m x n)}

\item{B}{Flattened matrix B (n x p)}

\item{m}{Rows of A}

\item{n}{Columns of A / Rows of B}

\item{p}{Columns of B}
}
\value{
Flattened product matrix (m x p)
}
\description{
Fast Matrix Multiplication (Row-Major flattened matrices)
}
\examples{
pytekt_matrix_multiply(c(1, 2, 3, 4), c(5, 6, 7, 8), 2, 2, 2)
}
