% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_max}
\alias{pytekt_max}
\title{Fast Maximum}
\usage{
pytekt_max(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Maximum value in vector
}
\description{
Fast Maximum
}
\examples{
pytekt_max(c(4.0, 1.0, 8.0, 2.0))
}
