% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_mean}
\alias{pytekt_mean}
\title{Fast Mean}
\usage{
pytekt_mean(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Arithmetic mean of vector elements
}
\description{
Fast Mean
}
\examples{
pytekt_mean(c(1.0, 2.0, 3.0, 4.0))
}
