% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_min}
\alias{pytekt_min}
\title{Fast Minimum}
\usage{
pytekt_min(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Minimum value in vector
}
\description{
Fast Minimum
}
\examples{
pytekt_min(c(4.0, 1.0, 8.0, 2.0))
}
