% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_norm2}
\alias{pytekt_norm2}
\title{Fast L2 Norm}
\usage{
pytekt_norm2(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Euclidean L2 norm
}
\description{
Fast L2 Norm
}
\examples{
pytekt_norm2(c(3.0, 4.0))
}
