% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_prefix_sum}
\alias{pytekt_prefix_sum}
\title{Fast Cumulative Prefix Sum}
\usage{
pytekt_prefix_sum(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Cumulative prefix sum vector
}
\description{
Fast Cumulative Prefix Sum
}
\examples{
pytekt_prefix_sum(c(1.0, 2.0, 3.0, 4.0))
}
