% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_rolling_mean}
\alias{pytekt_rolling_mean}
\title{Fast Rolling Mean}
\usage{
pytekt_rolling_mean(x, window)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Window size}
}
\value{
Rolling means
}
\description{
Fast Rolling Mean
}
\examples{
pytekt_rolling_mean(c(1.0, 2.0, 3.0, 4.0, 5.0), window = 3)
}
