% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_rolling_sum}
\alias{pytekt_rolling_sum}
\title{Fast Rolling Sum}
\usage{
pytekt_rolling_sum(x, window)
}
\arguments{
\item{x}{Numeric vector}

\item{window}{Window size}
}
\value{
Rolling sums
}
\description{
Fast Rolling Sum
}
\examples{
pytekt_rolling_sum(c(1.0, 2.0, 3.0, 4.0, 5.0), window = 3)
}
