% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_std}
\alias{pytekt_std}
\title{Fast Standard Deviation}
\usage{
pytekt_std(x, ddof = 0L)
}
\arguments{
\item{x}{Numeric vector}

\item{ddof}{Delta degrees of freedom (default: 0)}
}
\value{
Standard deviation of the vector
}
\description{
Fast Standard Deviation
}
\examples{
pytekt_std(c(1.0, 2.0, 3.0, 4.0), ddof = 1)
}
