% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_sum}
\alias{pytekt_sum}
\title{Fast Sum of Numeric Vector}
\usage{
pytekt_sum(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Sum of the vector elements
}
\description{
Fast Sum of Numeric Vector
}
\examples{
pytekt_sum(c(1.0, 2.0, 3.0, 4.0))
}
