% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pytekt_variance}
\alias{pytekt_variance}
\title{Fast Variance}
\usage{
pytekt_variance(x, ddof = 0L)
}
\arguments{
\item{x}{Numeric vector}

\item{ddof}{Delta degrees of freedom (default: 0)}
}
\value{
Variance of the vector
}
\description{
Fast Variance
}
\examples{
pytekt_variance(c(1.0, 2.0, 3.0, 4.0), ddof = 1)
}
