% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reticulate.R
\name{pytekt_version}
\alias{pytekt_version}
\title{Get Pytekt Python Module Version}
\usage{
pytekt_version()
}
\value{
Character string representing the Python package version, or NA if unavailable.
}
\description{
Get Pytekt Python Module Version
}
\examples{
\dontrun{
if (pytekt_available()) {
  pytekt_version()
}
}
}
